+++ reference/separan/builtins.py
@@ -44,6 +44,7 @@ from .http_server import SERVER_BUILTINS
 from .database import DB_BUILTINS
 from .collection_ops import average, count as count_values, filter_list, flatten, map_list, reduce_list, sum_list
 from .math_ops import MATH_BUILTINS
+from .mail import MAIL_BUILTINS
 
 
 MAX_TEXT_LENGTH = 1_048_576
@@ -787,8 +788,9 @@ BUILTINS = {
             "parse_error", "import_error", "regex_error", "glob_error", "argument_error", "permission_error"
             , "auth_error", "secret_error", "oauth_error"
             , "crypto_error", "crypto_authentication_error"
+            , "mail_error", "mail_address_error", "mail_attachment_error", "mail_provider_error", "mail_connection_error", "mail_authentication_error", "mail_send_error"
             , "cookie_error"
             , "db_connection_error", "db_auth_error", "db_query_error", "db_constraint_error", "db_timeout_error", "db_transaction_error", "db_driver_error"
         )),
-    ) + MATH_BUILTINS + (UtilityFunction("datetime", 1, 6, _datetime_construct, ("timezone",)),) + UTILITY_BUILTINS + PROCESS_BUILTINS + HTTP_BUILTINS + BYTES_BUILTINS + AUTH_BUILTINS + CRYPTO_BUILTINS + COOKIE_BUILTINS + COOKIE_STORE_BUILTINS + SERVER_BUILTINS + DB_BUILTINS
+    ) + MATH_BUILTINS + (UtilityFunction("datetime", 1, 6, _datetime_construct, ("timezone",)),) + UTILITY_BUILTINS + PROCESS_BUILTINS + HTTP_BUILTINS + BYTES_BUILTINS + AUTH_BUILTINS + CRYPTO_BUILTINS + MAIL_BUILTINS + COOKIE_BUILTINS + COOKIE_STORE_BUILTINS + SERVER_BUILTINS + DB_BUILTINS
 }